Improve badges’s alt text
For accessibility reasons, the HTML Standard has an entire section
laying out the requirements for alt text [1]. This change makes the
badges’s alt text more closely conform to the HTML Standard.

Specifically, the standard says:
    One way to think of alternative text is to think about how you would
    read the page containing the image to someone over the phone,
    without mentioning that there is an image present. Whatever you say
    instead of the image is typically a good start for writing the
    alternative text. (Source: [2])
To that end, this change turns the alt text into complete sentences.

Additionally, the standard goes on to give specific scenarios and what
should be done in those scenarios. These two scenarios are the most
applicable: [3] and [4]. Unfortunately, we don’t know what the images
are going to say, so we can’t follow [4].

[1]: <https://html.spec.whatwg.org/dev/images.html#alt>
[2]: <https://html.spec.whatwg.org/dev/images.html#general-guidelines>
[3]: <https://html.spec.whatwg.org/dev/images.html#a-link-or-button-containing-nothing-but-the-image>
[4]: <https://html.spec.whatwg.org/dev/images.html#text-that-has-been-rendered-to-a-graphic-for-typographical-effect>

Signed-off-by: Jason Yundt <jason@jasonyundt.email>