Clarify naming requirements for Sourcehut CI
Before this change, if a fork of this repo had a name other than
“radicle-link”, then Sourcehut CI wouldn’t work properly. It would clone
both the original repo and the fork, then proceed to run tests on the
original, not the fork.

This change makes it clear that your fork’s name has to be
“radicle-link”. It also makes sure that Sourcehut doesn’t clone the
original repo and that Sourcehut fails early in the event that a
different name is used.

Signed-off-by: Jason Yundt <jason@jasonyundt.email>